Add App Model

+++ MonkeyShelter.App/ShelterApp.cs
@@ -1,6 +1,7 @@
-﻿using System.Collections.Generic;
+﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
+using MonkeyShelter.App.Model;
 using MonkeyShelter.Data.Model;
 
 namespace MonkeyShelter.App
@@ -14,39 +15,74 @@ namespace MonkeyShelter.App
             _repository = repository;
         }
 
-        public async Task<IReadOnlyCollection<Monkey>> ListRegistryAsync(CancellationToken cancellationToken = default)
+        public async Task<MonkeyRegistry> GetRegistryAsync(CancellationToken cancellationToken = default)
         {
-            return await _repository.GetAllAsync(cancellationToken);
+            var monkeys = await _repository.GetAllAsync(cancellationToken);
+            return new MonkeyRegistry(monkeys);
         }
 
-        public async Task<Monkey> GetMonkeyDetailsAsync(string id, CancellationToken cancellationToken = default)
+        public async Task<MonkeyDetails> GetMonkeyDetailsAsync(string id, CancellationToken cancellationToken = default)
         {
+            if (id == null) throw new ArgumentNullException(nameof(id));
             var monkey = await _repository.GetByIdAsync(id, cancellationToken);
             if (monkey == null)
             {
                 throw new NotFoundException(id);
             }
-            return monkey;
+            return new MonkeyDetails(monkey);
         }
 
-        public async Task RegisterMonkeyAsync(Monkey monkey, CancellationToken cancellationToken = default)
+        public async Task RegisterMonkeyAsync(RegisterMonkeyRequest request, CancellationToken cancellationToken = default)
         {
+            if (request == null) throw new ArgumentNullException(nameof(request));
+            if (request.Id == null) throw new ArgumentException("Id null", nameof(request));
+            
+            var monkey = new Monkey()
+            {
+                Id = request.Id,
+                Name = request.Name,
+                Age = request.Age,
+                Weight = request.Weight,
+                EyeColor = request.EyeColor,
+                Species = request.Species,
+                Registered = request.Registered,
+                FavoriteFruit = request.FavoriteFruit,
+            };
+
             if (!await _repository.AddAsync(monkey, cancellationToken))
             {
-                throw new ConflictException(monkey.Id);
+                throw new ConflictException(request.Id);
             }
         }
 
-        public async Task UpdateRegistryAsync(Monkey monkey, CancellationToken cancellationToken = default)
+        public async Task UpdateRegistryAsync(UpdateRegistryRequest request, CancellationToken cancellationToken = default)
         {
+            if (request == null) throw new ArgumentNullException(nameof(request));
+            if (request.Id == null) throw new ArgumentException("Id null", nameof(request));
+
+            var monkey = await _repository.GetByIdAsync(request.Id, cancellationToken);
+            if (monkey == null)
+            {
+                throw new NotFoundException(request.Id);
+            }
+
+            monkey.Name = request.Name;
+            monkey.Age = request.Age;
+            monkey.Weight = request.Weight;
+            monkey.EyeColor = request.EyeColor;
+            monkey.Species = request.Species;
+            monkey.Registered = request.Registered;
+            monkey.FavoriteFruit = request.FavoriteFruit;
+
             if (!await _repository.UpdateAsync(monkey, cancellationToken))
             {
-                throw new NotFoundException(monkey.Id);
+                throw new NotFoundException(request.Id);
             }
         }
 
         public async Task ReleaseMonkeyAsync(string id, CancellationToken cancellationToken = default)
         {
+            if (id == null) throw new ArgumentNullException(nameof(id));
             if (!await _repository.RemoveByIdAsync(id, cancellationToken))
             {
                 throw new NotFoundException(id);
